# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Finna FinOps API
  version: 0.2.0
  description: 'The Finna FinOps API is the backend for the Finna cloud cost management platform. It handles multi-cloud cost
    ingestion (Azure Cost Management, GCP Billing, LLM usage), resource wastage detection, alert management, and per-project
    FinOps dashboards. Extracted cost records are normalised into a unified schema and made available through a rich query
    layer with time-window shortcuts (`mtd`, `7d`, `30d`, `90d`), provider filters, and CSV export.


    Authentication supports three flows: username/password JWT (primary), GitHub OAuth (developer login), and pluggable OIDC
    providers with PKCE and auto-provisioning. All resource endpoints require a valid JWT Bearer token. Admin-only endpoints
    (auth provider CRUD) require an `is_admin` claim in the token.


    This repository contains the API-only backend. The React/TypeScript UI lives in a separate repo and communicates exclusively
    through this API.'
  contact:
    name: Finna Team
    email: andrea.carmisciano@abstract.it
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /api/v1/auth/oidc/login:
    post:
      tags:
      - OIDC
      summary: Initiate OIDC login with PKCE
      description: Generates a PKCE code challenge, state, and nonce, then returns the provider's authorization URL. State
        is stored server-side for callback validation. Rate-limited per IP.
      operationId: oidc_login_api_v1_auth_oidc_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OIDCLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OIDCLoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/oidc/callback:
    post:
      tags:
      - OIDC
      summary: Complete OIDC callback and issue JWT
      description: Validates the OIDC state, exchanges the authorization code for tokens, verifies the ID token, applies claim
        mappings, and provisions or updates the local user. Issues a Finna JWT on success. Auto-provisioning is governed by
        the provider config flag.
      operationId: oidc_callback_api_v1_auth_oidc_callback_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OIDCCallbackRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OIDCCallbackResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OIDCCallbackRequest:
      properties:
        provider_id:
          type: string
          title: Provider Id
        code:
          type: string
          title: Code
        state:
          type: string
          title: State
      type: object
      required:
      - provider_id
      - code
      - state
      title: OIDCCallbackRequest
      description: Callback request.
    OIDCCallbackResponse:
      properties:
        token:
          type: string
          title: Token
        user_id:
          type: integer
          title: User Id
        username:
          type: string
          title: Username
      type: object
      required:
      - token
      - user_id
      - username
      title: OIDCCallbackResponse
      description: Callback response.
    OIDCLoginRequest:
      properties:
        provider_id:
          type: string
          title: Provider Id
      type: object
      required:
      - provider_id
      title: OIDCLoginRequest
      description: Login request.
    OIDCLoginResponse:
      properties:
        authorization_url:
          type: string
          title: Authorization Url
        state:
          type: string
          title: State
      type: object
      required:
      - authorization_url
      - state
      title: OIDCLoginResponse
      description: Login response.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/token
servers:
- url: http://localhost:8000
  description: Local development server
...